[vhost] Move endpoint ownership from TExecutor to TServer
Preparation for serving a single vhost endpoint by several executors. No
behaviour changes.

* Move the endpoint map, endpoint and device handler creation from
  TExecutor to TServer. TExecutor is now just a request queue plus the
  thread that runs it.
* PickExecutor keeps choosing the executor with the smallest number of
  assigned vhost queues, but counts them over TServer::Endpoints instead
  of a per-executor map.
* server.cpp had grown to hold every class of the vhost server. Move them
into separate translation units.
